requirements.txt
python-decouple==3.6
fastapi==0.86.0
fastjsonschema==2.15.3
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
emoji==1.7.0
et-xmlfile==1.1.0
idna==3.3
numpy==1.22.4
oauthlib==3.2.0
openai==0.19.0
openpyxl==3.0.10
pandas==1.4.2
python-crontab==2.6.0
pandas-stubs==1.2.0.61
python-dateutil==2.8.2
pytz==2022.1
regex==2022.6.2
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
tweepy==4.10.0
urllib3==1.26.9
uvicorn==0.19.0